Add employees resource on frontend and fix main issues on backend

+++ api/src/controllers/employee.controller.js
@@ -4,12 +4,14 @@ import {
     createEmployeeRequest,
     deleteEmployeeRequest,
     findEmployeeByIdRequest,
+    getAllEmployeesRequest,
     updateEmployeeRequest,
-} from "../validators/employee.validator.js";
+} from "@itedya/sor-tokajuk-pracownicy-ajax-validation";
 import db from "../database/knex.js";
+import ValidationHttpException from "../exceptions/validation.http-exception.js";
 
 const get = errorCatcherMiddleware(async (req, res, next) => {
-    const data = await findEmployeeByIdRequest.validate(req.body);
+    const data = await findEmployeeByIdRequest.validate(req.query);
 
     if (data.id) {
         await findEmployeeById(req, res, next, data.id);
@@ -19,19 +21,25 @@ const get = errorCatcherMiddleware(async (req, res, next) => {
 });
 
 const getEmployees = async (req, res, next) => {
-    const employees = await employeeRepository.get(db);
+    const data = await getAllEmployeesRequest.validate(req.query);
+
+    const employees = await employeeRepository.get(db, data.boss, data.team);
 
     res.status(200).json(employees.map((employee) => employee.sanitize()));
 };
 
 const findEmployeeById = async (req, res, next, id) => {
-    const employee = await employeeRepository.findById(db, employee.id);
+    const employee = await employeeRepository.findById(db, id);
 
     res.status(200).json(employee.sanitize());
 };
 
-const createEmployee = async (req, res, next) => {
+const createEmployee = errorCatcherMiddleware(async (req, res, next) => {
     const data = await createEmployeeRequest.validate(req.body);
+
+    data.employeedFrom = new Date(data.employeedFrom)
+        .toISOString()
+        .slice(0, 10);
 
     const employee = await db.transaction((trx) => {
         return employeeRepository.create(
@@ -48,9 +56,9 @@ const createEmployee = async (req, res, next) => {
     });
 
     res.status(200).json(employee.sanitize());
-};
+});
 
-const updateEmployee = async (req, res, next) => {
+const updateEmployee = errorCatcherMiddleware(async (req, res, next) => {
     const data = await updateEmployeeRequest.validate(req.body);
 
     const employee = await db.transaction((trx) => {
@@ -69,17 +77,21 @@ const updateEmployee = async (req, res, next) => {
     });
 
     res.status(200).json(employee.sanitize());
-};
+});
 
-const deleteEmployee = async (req, res, next) => {
+const deleteEmployee = errorCatcherMiddleware(async (req, res, next) => {
     const data = await deleteEmployeeRequest.validate(req.body);
 
-    await db.transaction((trx) => {
+    await db.transaction(async (trx) => {
+        if ((await employeeRepository.findById(trx, data.id)) === null) {
+            throw new ValidationHttpException(["Podane ID jest niepoprawne."]);
+        }
+
         return employeeRepository.delete(trx, data.id);
     });
 
     res.status(200).send();
-};
+});
 
 const employeeController = {
     get,
